Part-way through making pause-menu buttons a consistent size. Limitation is, calculateTextSize() crashes if the text has to wrap because of the way I originally implemented wrapping. That has to change! CLOC: 9352

+++ src/ui_window.cpp
@@ -85,7 +85,7 @@ bool window_button(WindowContext *context, String text, s32 textWidth=-1)
 		}
 		else
 		{
-			maxWidth = textWidth;
+			maxWidth = (f32) textWidth;
 		}
 
 		BitmapFontCachedText *textCache = drawTextToCache(context->temporaryMemory, font, text, style->textColor, maxWidth);
